Module bitgrin_wallet_libwallet::api_impl::owner

Generic implementation of owner API functions

Functions

accounts
List of accounts

cancel_tx
cancel tx

check_repair
check repair

create_account_path
new account path

finalize_tx
Finalize slate

get_stored_tx
get stored tx

init_send_tx
Initiate tx as sender

issue_invoice_tx
Initiate a transaction as the recipient (invoicing)

node_height
node height

post_tx
Posts a transaction to the chain take a client impl instead of wallet so as not to have to lock the wallet

process_invoice_tx
Receive an invoice tx, essentially adding inputs to whatever output was specified

restore
Attempt to restore contents of wallet

retrieve_outputs
retrieve outputs

retrieve_summary_info
Retrieve summary info

retrieve_txs
Retrieve txs

set_active_account
set active account

tx_lock_outputs
Lock sender outputs

verify_slate_messages
verify slate messages